#189 Multiple file changes: Created output.py Renamed: core_calculations -> core_computations,     files -> local_files Moved prepare_partridge_feed, save_trip_stats, save_route_stats, remote_key_to_local_path
